COD payment provider plugin
the payment step in checkout workflow.

There's an additional shortcoming when it comes to order processing. The
reason is that a order can't be kept in processing once it gets shipped.

But during the nature of COD, the payment will occur after the order has
been shipped. Therefore it should still be in mode : "processing", which
is not possible ATM.